PROTOTYPE: tier-2 period-detecting caterpillar + comparison benchmark
Adds Segment::Periodic + gated period detection (prefix-recurrence gate,
budget floor, Booth least-rotation canonicalization for phase-independent
dedup identity), and examples/compare.rs benchmarking fastcdc-v2020 (4.0.1)
vs plain mincdc vs cat-simple (tier 1) vs cat-period (tier 2).

Key measured findings (16 MiB inputs, ~8 KiB target, native arm64/NEON):

  mincdc vs fastcdc: mincdc ~5x faster on normal data (9.1 vs 1.7 GB/s),
  predictable ~target chunk sizes vs fastcdc's long tail (zero-fill: fastcdc
  forced 51 KiB chunks). Versioned dedup ~equal (49.7% vs 49.9%) but mincdc
  uses far fewer records (567 vs 1682).

  cat-simple (tier 1) is the clear winner: free on random/text (within noise),
  collapses metadata hard where it can -- zero-fill 8192->1, periodic-wide
  7198->3, mixed 4091->1154 -- and never hurts dedup.

  cat-period (tier 2) does NOT earn its keep as a default:
   * mincdc SELF-ALIGNS boundaries to periods under normal wide windows, so
     tier 1 already gets the win; tier 2 only adds a marginal 3->1.
   * its unique benefit (7853->1) appears only in NARROW windows where no
     period multiple fits -- an unusual config.
   * the gate runs on every non-coalesced chunk, costing a large throughput
     hit (random 9->2.5, text 3.4->1.1, mixed 4.1->0.6, narrow 42->2.2 GB/s).
  Conclusion: ship tier 1; keep tier 2 as a narrow-window-only / research option.
  (The gate is unoptimized; a memchr-based probe would cut the random penalty,
  but cannot beat tier 1, which is free and already captures the wide-window win.)

Bumps fastcdc dev-dep 3.2.1 -> 4.0.1 (usize size args; updated examples).

+++ examples/chunk_only.rs
@@ -30,9 +30,9 @@ fn chunk_slice(bytes: &[u8], alg: &str, min_chunk_size: usize, mean_chunk_size: 
             let max_chunk_size = mean_chunk_size + (mean_chunk_size - min_chunk_size) * 7;
             let mut chunker = fastcdc::v2020::FastCDC::new(
                 bytes,
-                min_chunk_size as u32,
-                mean_chunk_size as u32,
-                max_chunk_size as u32,
+                min_chunk_size,
+                mean_chunk_size,
+                max_chunk_size,
             );
             while let Some(chunk) = chunker.next() {
                 std::hint::black_box(bytes[chunk.offset]);
